Use lua_State from loop:loop() call site, not watcher creation
The lua_State for running callbacks used to be stored in a thread-local
variable initialized when loop:loop() was called.  At some point this
changed to be a lua_State that is stored (in watcher->data) when a
watcher is created.  This is almost always the wrong thing in an
environment with coroutines.

Instead, abuse the userdata slot in the loop object for this.  Save
off the old userdata before calling ev_loop(), set it to L, run ev_loop(),
then restore it.  This gets roughly the same effect as the old code
without requiring any pthreads nastiness, since if you're running the
same loop from two different threads at the same time you've already
got a problem.