Save the lua state differently in watchers
This code used to use ev_userdata() for saving the lua_State*. However, I happen
to use this library inside of some other code and that other code handles the
loop. This means that ev_userdata() was never set up and this code ran into a
NULL pointer dereference.

To fix this, this commits now uses each watcher's data member to save the
lua_State*. This also makes it impossible to clash with other code which tries
to save some user data.

Signed-off-by: Uli Schlachter <psychon@znc.in>